refactored `AddUserToClientCommandHandler`
