+++ ReportHub/ReportHub.Application/Contracts/IdentityContracts/IIdentityService.cs
@@ -2,6 +2,8 @@
 
 public interface IIdentityService
 {
+    Task<bool> AssignUserRole(string userId, string roleName);
+
     Task<bool> ValidateUserIdExists(string userId, CancellationToken cancellationToken);
 
     Task<bool> ValidateRoleExists(string role, CancellationToken cancellationToken);
