+++ ReportHub/ReportHub.Infrastructure/Services/IdentityServices/IdentityService.cs
@@ -1,6 +1,9 @@
 ﻿using Microsoft.AspNetCore.Http;
 using ReportHub.Application.Contracts.IdentityContracts;
+using ReportHub.Domain.Entities;
 using System.Net.Http.Headers;
+using System.Net.Http.Json;
+using System.Threading;
 
 namespace ReportHub.Infrastructure.Services.IdentityServices;
 
@@ -20,6 +23,18 @@ public class IdentityService : IIdentityService
         {
             _httpClient.DefaultRequestHeaders.Authorization = new AuthenticationHeaderValue("Bearer", token);
         }
+    }
+
+    public async Task<bool> AssignUserRole(string userId, string roleName, CancellationToken cancellationToken)
+    {
+        var requestBody = new
+        {
+            userId = userId,
+            roleName = roleName
+        };
+        var result = await _httpClient.PostAsJsonAsync("https://localhost:7171/api/Admin/assign-role", requestBody, cancellationToken);
+
+        return result.IsSuccessStatusCode;
     }
 
     public async Task<bool> ValidateUserIdExists(string userId, CancellationToken cancellationToken)
